mcp_server/requirements.txt
mcp>=1.1.0
pydantic>=2.0.0
python-dateutil==2.8.2
pytest-asyncio